Simple ansible playbook that illustrates a problem with conditional role dependencies.
Role a depends on role b when var want_b is true.
Role b depends on role c when var want_c is true.

When want_b is false and want_c is true I expect both
roles b and c to be skipped, because role b is skipped
and role c is a dependency of role b.